+++ R/krona.R
@@ -1,7 +1,11 @@
 ################################################################################
 #' Make Krona files using [KronaTools](https://github.com/marbl/Krona/wiki).
+#' @description 
+#'`r lifecycle::badge("maturing")`
+#' 
 #' Need the installation of kronatools on the computer ([installation instruction](https://github.com/marbl/Krona/wiki/Installing)).
 #' Function merge_krona allows merging multiple html files in one interactive krona file
+#' 
 #' @aliases merge_krona
 #' @param physeq (required): a \code{\link{phyloseq-class}} object.
 #' @param file (required): the location of the html file to save