+++ NEWS.md
@@ -1,3 +1,9 @@
+# MiscMetabar 0.22
+
+* Add badge to set the development lifecycle of each function
+* Add the lulu_phyloseq function to make easy the reclustering of phyloseq object using the lulu algorithm (https://www.nature.com/articles/s41467-017-01312-x) from the [lulu package](https://github.com/adrientaudiere/lulu).
+
+
 # MiscMetabar 0.21
 
 * This is the first release of pkgdown.